Restructuring the package.
