+++ omni_wagtail_library/__init__.py
@@ -5,4 +5,4 @@ Page library for Wagtail
 from __future__ import unicode_literals
 
 
-__version__ = '0.0.4'
+__version__ = '0.1.0'
