+++ omni_wagtail_library/models.py
@@ -5,91 +5,16 @@ Application models
 
 from __future__ import unicode_literals
 
-from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.db import models
-
 from wagtail.wagtailadmin.edit_handlers import FieldPanel, StreamFieldPanel
-from wagtail.wagtailcore.blocks import RichTextBlock, PageChooserBlock
+from wagtail.wagtailcore.blocks import RichTextBlock
 from wagtail.wagtailcore.fields import RichTextField, StreamField
 from wagtail.wagtailcore.models import Page
 from wagtail.wagtailimages.blocks import ImageChooserBlock
 
-
-class AbstractLibraryListingPage(Page):
-    """
-    Abstract library listing page
-    """
-    paginate_by = models.PositiveIntegerField(
-        blank=True,
-        null=True
-    )
-
-    content_panels = Page.content_panels + [
-        FieldPanel('paginate_by'),
-    ]
-
-    class Meta(object):
-        """
-        Django model meta options
-        """
-        abstract = True
-
-    def _get_children(self, request):
-        """
-        Helper method for getting child nodes to display in the listing
-        :param request: django request
-        :return: Queryset of child model instances
-        """
-        model_class = self.__class__.allowed_subpage_models()[0]
-        children = model_class.objects.child_of(self)
-        if not request.is_preview:
-            children = children.filter(live=True)
-        return children
-
-    def _paginate_queryset(self, queryset, page):
-        """
-        Helper method for paginating the queryset provided
-        :param queryset: Queryset of model instances to paginate
-        :param page: Raw page number taken from the request dict
-        :return: Queryset of child model instances
-        """
-        paginator = Paginator(queryset, self.paginate_by)
-        try:
-            queryset = paginator.page(page)
-        except PageNotAnInteger:
-            queryset = paginator.page(1)
-        except EmptyPage:
-            queryset = paginator.page(paginator.num_pages)
-        return queryset, paginator
-
-    def get_context(self, request, *args, **kwargs):
-        """
-        Adds child pages to the context and paginates them if pagination is required
-        :param request: HttpRequest instance
-        :param args: default positional args
-        :param kwargs: default keyword args
-        :return: Context data to use when rendering the template
-        """
-        context = super(AbstractLibraryListingPage, self).get_context(request, *args, **kwargs)
-        queryset = children = self._get_children(request)
-        is_paginated = False
-        paginator = None
-
-        # Paginate the child nodes if paginate_by has been specified
-        if self.paginate_by:
-            is_paginated = True
-            children, paginator = self._paginate_queryset(children, request.GET.get('page'))
-
-        context.update(
-            queryset=queryset,
-            children=children,
-            paginator=paginator,
-            is_paginated=is_paginated
-        )
-        return context
+from omni_wagtail_library import abstract_models
 
 
-class LibraryListingPage(AbstractLibraryListingPage):
+class LibraryListingPage(abstract_models.AbstractLibraryListingPage):
     """
     Common features
     """
@@ -102,24 +27,7 @@ class LibraryListingPage(AbstractLibraryListingPage):
     subpage_types = ['omni_wagtail_library.LibraryItemDetailPage']
 
 
-class AbstractLibraryItemDetailPage(Page):
-    """
-    Abstract library item detail page
-    """
-
-    attachment = models.FileField(upload_to='attachments')
-    content_panels = Page.content_panels + [
-        FieldPanel('attachment'),
-    ]
-
-    class Meta(object):
-        """
-        Django properties
-        """
-        abstract = True
-
-
-class LibraryItemDetailPage(AbstractLibraryItemDetailPage):
+class LibraryItemDetailPage(abstract_models.AbstractLibraryItemDetailPage):
     """
     Library item detail page
     """
@@ -132,23 +40,3 @@ class LibraryItemDetailPage(AbstractLibraryItemDetailPage):
         StreamFieldPanel('content'),
         FieldPanel('attachment'),
     ]
-
-
-class LibraryItemBlock(PageChooserBlock):
-    """
-    Library Item chooser block for streamfield
-    """
-    def __init__(self, target_model='omni_wagtail_library.LibraryItemDetailPage', **kwargs):
-        """
-        Initialization code
-
-        :param target_model: default selection model - can be overriden in model
-        :param kwargs: default block kwargs
-        """
-        super(LibraryItemBlock, self).__init__(target_model, **kwargs)
-
-    class Meta(object):
-        """
-        Block meta
-        """
-        template = 'omni_wagtail_library/library_item_block.html'
